AsarVersioner: Fix for Header Parsing Issue
After some debugging, found that the header size was off just enough to add a null character to the end of the string. Rather than trying to recalculate it every time, I just stripped it off. I may need to make this smarter in the future, but this is a reasonable brute force method.

Also added a more verbose error message.