Fixes for ReadWrite7.licensed.pkg
Added new global input for `APP_NAME`

Switched from `AppDmgVersioner` to `PlistReader` since the app is now downloaded as a ZipFile and not a DMG and the Processor was failing to read the file contents. Since the parent recipe unzips the app to get the version number, this recipe uses that version for getting the version number and `bundleid` for `AppPkgCreator`.